Move case-for-support page to app router
